// src/modules/fairDb/dto/prodileDb.util.dto.ts
export class DynamicBmOtherEntry {
    fairRegistrationId: string;
    formFieldId: string | null;
    labelEn: string | null;
    labelTc: string | null;
    labelSc: string | null;
    value: string | null;
    valueEn: string | null;
    valueTc: string | null;
    valueSc: string | null;
    createdBy: string;
    lastUpdatedBy: string;
}